+++ frontend/src/app/(public)/profile/NavTabs/PersonalInfo/PersonalInfo.js
@@ -6,8 +6,8 @@ import { FaEdit, FaSave, FaUndo, FaUser } from 'react-icons/fa';
 import { OptimizedImage } from '@/components';
 import CustomSelect from './CustomSelect/CustomSelect';
 import DeleteAccount from './DeleteAccount/DeleteAccount';
-import { useFormValidation } from '../../hooks/useFormValidation';
-import { useProfileApi } from '../../hooks/useApiCall';
+import { useFormValidation } from '@/hooks/(public)/useFormValidation';
+import { useProfileApi } from '@/hooks/(public)/useApiCall';
 import { useToast } from '../../../components/Toast';
 import { formatBirthDate, formatDateForInput } from '@/utils/shared/dateUtils';
 import { getProfilePhotoUrl } from '@/utils/shared/uploadPaths';
