--- frontend/src/app/admin/organization/utils/index.js
@@ -1,2 +1,0 @@
-export { PhotoUtils } from './photoUtils';
-export { applyRoleHierarchyOrdering, sortHeadsByOrder, filterHeads, ROLE_OPTIONS } from './roleHierarchy';
