+++ frontend/src/hooks/(public)/useFormValidation.js
@@ -77,3 +77,4 @@ export const useFormValidation = () => {
     setFieldError
   };
 };
+
