+++ frontend/src/utils/admin/organizationUtils.js
@@ -1,0 +1,3 @@
+export { PhotoUtils } from './photoUtils';
+export { applyRoleHierarchyOrdering, sortHeadsByOrder, filterHeads, ROLE_OPTIONS } from './roleHierarchy';
+
